Refactor: Improve router password tester accuracy and reliability
