--- README_FIXES.md
@@ -1,112 +1,0 @@
-# مشکلات اصلی کد شما و راه‌حل‌ها
-
-## 🔍 مشکلات شناسایی شده:
-
-### 1. **مشکل اصلی: منطق تشخیص صفحه مدیریت**
-**مشکل:** سیستم امتیازبندی خیلی سخت‌گیرانه بود و معیارهای نادرست داشت
-**راه‌حل:**
-- بهبود سیستم امتیازبندی با وزن‌های منطقی‌تر
-- اضافه کردن بررسی چندمرحله‌ای
-- بهبود تشخیص عناصر منفی (مثل فیلد پسورد)
-
-### 2. **مشکل زمان‌بندی و انتظار**
-**مشکل:** زمان‌های انتظار نامناسب و عدم مدیریت درست بارگذاری صفحه
-**راه‌حل:**
-- بهبود `wait_for_page_load()` با چک‌های چندگانه
-- افزایش زمان انتظار پس از ارسال فرم
-- مدیریت بهتر محتوای پویا
-
-### 3. **مشکل تشخیص عناصر ورود**
-**مشکل:** تشخیص ناکافی فیلدهای نام کاربری و پسورد
-**راه‌حل:**
-- اضافه کردن selectors بیشتر برای فیلدهای ورود
-- پشتیبانی از فرمت‌های مختلف فرم
-- بهبود تشخیص دکمه ورود
-
-### 4. **مشکل مدیریت popup و alert**
-**مشکل:** مدیریت ناکافی popup ها و alert های JavaScript
-**راه‌حل:**
-- بهبود `handle_alerts_and_popups()`
-- اضافه کردن timeout مناسب
-- مدیریت انواع مختلف popup
-
-### 5. **مشکل HTTP testing**
-**مشکل:** تست HTTP خیلی ساده و ناکافی بود
-**راه‌حل:**
-- اضافه کردن فرمت‌های مختلف login data
-- بهبود منطق تشخیص موفقیت
-- افزایش timeout
-
-## 🚀 بهبودهای اعمال شده:
-
-### ✅ سیستم امتیازبندی جدید:
-```python
-strong_indicators = {
-    'logout': 15, 'log out': 15, 'sign out': 15,
-    'dashboard': 12, 'administration': 12, 'admin panel': 12,
-    'management': 10, 'configuration': 10, 'settings': 10,
-    'status': 8, 'system info': 8, 'device info': 8
-}
-
-medium_indicators = {
-    'wireless': 6, 'network': 6, 'wan': 6, 'lan': 6,
-    'firewall': 5, 'nat': 5, 'dhcp': 5, 'qos': 5,
-    'port forwarding': 5, 'access control': 5,
-    'firmware': 4, 'backup': 4, 'restore': 4
-}
-
-negative_indicators = {
-    'password': -3, 'login': -3, 'sign in': -3,
-    'username': -2, 'user name': -2, 'enter password': -4,
-    'forgot password': -5, 'remember me': -2
-}
-```
-
-### ✅ منطق تصمیم‌گیری بهبود یافته:
-```python
-if score >= 20:
-    return True, f"High confidence (score: {score})"
-elif score >= 15 and current_url != original_url:
-    return True, f"Good confidence with URL change (score: {score})"
-elif score >= 10 and current_title != original_title:
-    return True, f"Moderate confidence with title change (score: {score})"
-elif score >= 8 and any('logout' in indicator or 'dashboard' in indicator for indicator in indicators):
-    return True, f"Strong indicators present (score: {score})"
-```
-
-### ✅ بهبود تشخیص عناصر:
-- اضافه کردن selectors بیشتر
-- پشتیبانی از فرمت‌های مختلف
-- بهبود error handling
-
-### ✅ بهبود عملکرد:
-- کاهش زمان‌های غیرضروری
-- بهینه‌سازی Chrome options
-- مدیریت بهتر منابع
-
-## 🧪 نحوه تست:
-
-```bash
-# تست عملکرد کلی
-python test_brute_forcer.py
-
-# تست روی یک هدف
-python smart_brute_forcer_fixed.py --target 192.168.1.1 --mode both
-
-# تست روی محدوده IP
-python smart_brute_forcer_fixed.py --target 192.168.1.1-192.168.1.10 --mode chrome
-```
-
-## 📊 تغییرات کلیدی:
-
-1. **سیستم امتیازبندی هوشمند** با وزن‌های منطقی
-2. **تشخیص چندمرحله‌ای** صفحه مدیریت
-3. **مدیریت بهتر زمان‌بندی** و انتظار
-4. **پشتیبانی از فرمت‌های مختلف** ورود
-5. **مدیریت کامل popup** و alert
-6. **بهبود error handling** و logging
-7. **افزایش لیست پسورد** با پسوردهای رایج
-8. **بهینه‌سازی عملکرد** کلی
-
-## 🎯 نتیجه:
-برنامه حالا باید بتواند پسوردها را با دقت بالاتری تشخیص دهد و مشکلات قبلی را نداشته باشد.
