--- README_SIMPLE.md
@@ -1,108 +1,0 @@
-# تغییرات اعمال شده طبق درخواست شما
-
-## 🎯 مشکلات حل شده:
-
-### ✅ 1. حذف تست یوزرنیم
-**مشکل:** برنامه یوزرنیم هم تست می‌کرد
-**راه‌حل:** 
-- حذف کامل تست یوزرنیم از هر دو روش HTTP و Chrome
-- فقط فیلد پسورد پر می‌شود
-- فیلد یوزرنیم نادیده گرفته می‌شود
-
-### ✅ 2. محدود کردن لیست پسورد
-**مشکل:** لیست پسورد خیلی طولانی بود
-**راه‌حل:**
-```python
-self.password_list = ["admin", "JAMES1", "admin1", "user"]
-```
-فقط همین 4 پسورد تست می‌شود
-
-### ✅ 3. توقف فوری پس از پیدا کردن پسورد
-**مشکل:** برنامه بعد از پیدا کردن پسورد درست، پسوردهای دیگر هم تست می‌کرد
-**راه‌حل:**
-- بلافاصله پس از موفقیت، تست متوقف می‌شود
-- پیام "🛑 STOPPING TESTS - Password found and verified!" نمایش داده می‌شود
-- هیچ پسورد دیگری تست نمی‌شود
-
-## 📁 فایل‌های ایجاد شده:
-
-### 1. `smart_brute_forcer_fixed.py` (نسخه اصلاح شده اصلی)
-- تمام مشکلات حل شده
-- سیستم امتیازبندی بهبود یافته
-- توقف فوری پس از موفقیت
-
-### 2. `simple_password_tester.py` (نسخه ساده و مخصوص)
-- فقط 4 پسورد مشخص شما
-- فقط تست پسورد (بدون یوزرنیم)
-- توقف فوری پس از موفقیت
-- کد ساده‌تر و قابل فهم‌تر
-
-### 3. `test_simple.py` (تست نسخه ساده)
-- تست عملکرد نسخه ساده
-
-## 🚀 نحوه استفاده:
-
-### نسخه ساده (پیشنهادی):
-```bash
-# تست با هر دو روش
-python simple_password_tester.py --target 192.168.1.1 --mode both
-
-# فقط HTTP (سریع‌تر)
-python simple_password_tester.py --target 192.168.1.1 --mode http
-
-# فقط Chrome (دقیق‌تر)
-python simple_password_tester.py --target 192.168.1.1 --mode chrome
-```
-
-### نسخه کامل:
-```bash
-python smart_brute_forcer_fixed.py --target 192.168.1.1 --mode both
-```
-
-### تست عملکرد:
-```bash
-python test_simple.py
-```
-
-## 🎯 ویژگی‌های نسخه ساده:
-
-1. **فقط 4 پسورد:** admin, JAMES1, admin1, user
-2. **بدون یوزرنیم:** فقط پسورد تست می‌شود
-3. **توقف فوری:** بلافاصله پس از پیدا کردن پسورد متوقف می‌شود
-4. **دو روش تست:** HTTP (سریع) و Chrome (دقیق)
-5. **خروجی واضح:** نتایج به صورت مفصل نمایش داده می‌شود
-
-## 📊 مثال خروجی:
-
-```
-🚀 SIMPLE PASSWORD TESTER
-========================================
-🎯 Target: 192.168.1.1
-🔐 Passwords to test: ['admin', 'JAMES1', 'admin1', 'user']
-
-[1/4] Testing password: admin
-   🌐 HTTP test...
-   ❌ HTTP Failed
-   🔍 Chrome test...
-   ❌ Chrome Failed
-
-[2/4] Testing password: JAMES1
-   🌐 HTTP test...
-   🎉 HTTP SUCCESS! Password 'JAMES1' works!
-🛑 Stopping tests - Password verified!
-
-🎉 SUCCESS!
-Target: 192.168.1.1
-Password: JAMES1
-Method: http
-Time: 2.3s
-```
-
-## 🔧 تنظیمات مهم:
-
-- **فقط پسورد:** یوزرنیم نادیده گرفته می‌شود
-- **توقف فوری:** پس از اولین موفقیت متوقف می‌شود
-- **4 پسورد مشخص:** طبق درخواست شما
-- **تأیید دوگانه:** هم پسورد و هم پنل مدیریت تأیید می‌شود
-
-حالا برنامه دقیقاً طبق نیاز شما کار می‌کند! 🎉
